% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicData.R
\docType{data}
\name{basicData}
\alias{basicData}
\title{A small basic dataset}
\format{A data frame with 5 rows and 2 columns
\describe{
  \item{basicDataX}{simulated X values}
  \item{basicDataY}{simulated Y values}
}}
\source{
I made this up
}
\usage{
basicData
}
\description{
This is just an example to see if I can get
this to work
}
\keyword{datasets}
